+++ docs/source/develop/changelog.rst
@@ -8,6 +8,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) because it
 
 
 
+[Unversioned] - Changes
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Changed
+---------------------
+
+- Gravity sensors attached to BrewPi controllers will now send those controller's temps to Brewfather
+
+
+
 [2019-12-15] - Brewer's Friend, Brewfather, and MacOS BLE Support
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~